Use Next 13.4 App Directory
