+++ .eslintrc.json
@@ -1,6 +1,3 @@
 {
-  "extends": "next/core-web-vitals",
-  "rules": {
-    "@next/next/no-img-element": 0
-  }
+  "extends": "next/core-web-vitals"
 }
