--- .prettierrc
@@ -1,9 +1,0 @@
-{
-	"trailingComma": "none",
-	"arrowParens": "avoid",
-	"singleQuote": true,
-	"printWidth": 80,
-	"useTabs": false,
-	"tabWidth": 2,
-	"semi": true
-}
