--- assets/languages.json
@@ -1,79 +1,0 @@
-{
-  "main1": [
-    "JavaScript",
-    "Python",
-    "C",
-    "C++",
-    "Rust",
-    "HTML",
-    "TypeScript",
-    "PHP",
-    "Dart",
-    "Java",
-    "Elixir",
-    "Objective-C",
-    "Swift",
-    "Kotlin",
-    "Vue",
-    "Svelte",
-    "Go",
-    "CSS",
-    "Ruby",
-    "CSharp",
-    "Shell",
-    "MATLAB",
-    "Bash",
-    "Scss",
-    "Haskell",
-    "R",
-    "Sql",
-    "Perl",
-    "Groovy",
-    "WebAssembly",
-    "Pascal",
-    "CoffeeScript",
-    "Crystal",
-    "Clojure"
-  ],
-
-  "main": [
-    "JavaScript",
-    "Python",
-    "C",
-    "C++",
-    "Rust",
-    "HTML",
-    "TypeScript",
-    "PHP",
-    "Dart",
-    "Java",
-    "Elixir",
-    "Ruby"
-  ],
-  "others": [
-    "Objective-C",
-    "Swift",
-    "Kotlin",
-    "Vue",
-    "Svelte",
-    "Go",
-    "CSS",
-    "Ruby",
-    "CSharp",
-    "Shell",
-    "Rust",
-    "MATLAB",
-    "Bash",
-    "Scss",
-    "Haskell",
-    "R",
-    "Sql",
-    "Perl",
-    "Groovy",
-    "WebAssembly",
-    "Pascal",
-    "CoffeeScript",
-    "Crystal",
-    "Clojure"
-  ]
-}
