+++ package.json
@@ -6,36 +6,20 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "format": "prettier --write \"./**/*.{ts,tsx,json}\""
+    "lint": "next lint"
   },
   "dependencies": {
-    "@fontsource/poppins": "^4.5.10",
-    "@fortawesome/fontawesome-svg-core": "^6.4.0",
-    "@fortawesome/free-regular-svg-icons": "^6.4.0",
-    "@fortawesome/free-solid-svg-icons": "^6.4.0",
-    "@fortawesome/react-fontawesome": "^0.2.0",
-    "@twuni/emojify": "^1.0.2",
-    "daisyui": "^2.51.6",
-    "framer-motion": "^10.12.7",
-    "next": ">=13.4.0",
-    "next-seo": "^6.0.0",
-    "nextjs-progressbar": "^0.0.16",
+    "@types/node": "20.1.1",
+    "@types/react": "18.2.6",
+    "@types/react-dom": "18.2.4",
+    "autoprefixer": "10.4.14",
+    "eslint": "8.40.0",
+    "eslint-config-next": "13.4.1",
+    "next": "13.4.1",
+    "postcss": "8.4.23",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.43.9",
-    "react-icons": "^4.8.0"
-  },
-  "devDependencies": {
-    "@types/node": "^18.16.3",
-    "@types/react": "18.2.5",
-    "@types/twuni__emojify": "^1.0.0",
-    "autoprefixer": "^10.4.14",
-    "eslint": "8.39.0",
-    "eslint-config-next": "13.4.0",
-    "postcss": "^8.4.23",
-    "prettier": "^2.8.8",
-    "tailwindcss": "^3.3.2",
+    "tailwindcss": "3.3.2",
     "typescript": "5.0.4"
   }
 }
