+++ src/app/page.tsx
@@ -1,0 +1,9 @@
+import Hero from '@/components/Hero';
+
+export default function Home() {
+  return (
+    <main>
+      <Hero />
+    </main>
+  );
+}
