+++ src/utils/capFirstLetter.ts
@@ -1,0 +1,7 @@
+const capFirstLetter = (str: string) => {
+  return decodeURIComponent(str).replace(/(^|\s)\S/g, letter =>
+    letter.toUpperCase()
+  );
+};
+
+export default capFirstLetter;
