--- utils/capFirstLetter.ts
@@ -1,5 +1,0 @@
-const capFirstLetter = (str: string) => {
-  return str.replace(/(^|\s)\S/g, letter => letter.toUpperCase());
-};
-
-export default capFirstLetter;
